refactor(update): isolate finalize command
